Update expenses query to handle amounts bigger than 1,000

+++ resources/queries/recent_expenses.sql
@@ -1,5 +1,5 @@
 -- Retrieve recent expenses
-SELECT id, "when", category, description, amount
+SELECT id, "when", category, description, amount::decimal
 FROM expenses
 ORDER BY "when" DESC, id DESC
 LIMIT :count
